Updated parts of the system to use Sprite plotting.
I'm trying to convert the code to work on RISC OS Pyromaniac in any
mode, which means making it use sprites for the rendering. This is
a little fiddly and I'm pretty certain I've got the colour handling
wrong here. But it's kinda working, so I'm committing this and moving
to the next stage of magic...